holi va queriendo

+++ database/migrations/2025_11_30_222521_add_role_and_theme_to_users_table.php
@@ -18,11 +18,15 @@ return new class extends Migration {
     }
 
     public function down(): void
-    {
-        Schema::table('users', function (Blueprint $table) {
-            // Para revertir: elimina las columnas
+{
+    Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasColumn('users', 'theme')) {
             $table->dropColumn('theme');
+        }
+        if (Schema::hasColumn('users', 'role')) {
             $table->dropColumn('role');
-        });
-    }
+        }
+    });
+}
+
 };
